fix BlockEggBasket and creating TileAutoMaker yet.

+++ mods/applemilk/common/PacketHandler.java
@@ -398,7 +398,7 @@ public class PacketHandler implements IPacketHandler
     		ByteArrayOutputStream bos = new ByteArrayOutputStream();
             DataOutputStream dos = new DataOutputStream(bos);
 
-            byte Mode = (byte) container.getMode();
+            byte Mode = container.getMode();
 
             try
             {
